Don't allocate temporary sized_buffer every `]`
This cuts down on the number of allocations massively. For example,
compiling colortest.bf for x86_64 without optimizations goes from 110
malloc calls of around 549 KiB total to 4 malloc calls of around 9 KiB
total.

With optimizations, it goes from 110 malloc calls of around 431 KiB to 6
malloc calls of around 15 KiB.

tested with eambfc compiled on and for Debian 12 amd64, with CFLAGS=-O0,
and numbers gathered with GNU memusage. Reallocs were identical before
and after this change, so were not factored in.