Version 3.1.0
The big change is the addition of a backend for Linux on IBM
z/Architecture mainframes, which is disabled by default but can be
compiled in optionally. This was to find any bugs that may come up when
writing backends for big-endian targets. One such issue was found and
fixed.

A few minor fixes to logic oversights are also included. These
oversights are not known to have caused any real-world bugs, but had the
potential to do so.

Some of the code was streamlined.