Port remaining applicable riscv64 tests from rust
Not all tests are applicable, due to implementation differences - tests
of failure states don't carry over, as error handling is different in
the Rust version, and the c_addi and addi helper functions don't exist
in the C version, and thus can't be tested.

test_add_sub_byte is currently failing.