Fix broken unit tests, and rework arch_inter
Functions in `arch_inter` are now annotated with `nonnull_ret`, their
parameters are now `restrict`, and fallible functions set error through
a pointer, allowing caller to profide more information before display.

While implementing that, the issue fixed in the previous commit was
found. The changes to the unit tests to fix issues that were hidden by
the EXPECT_BF_ERR macro and the changes that were done to account for
the `jump_open` and `jump_close` changes could not be easily separated.